Use size_t instead of int.

+++ src/options.h
@@ -21,8 +21,8 @@ typedef struct {
     int ackmate;
     pcre *ackmate_dir_filter;
     pcre_extra *ackmate_dir_filter_extra;
-    int after;
-    int before;
+    size_t after;
+    size_t before;
     enum case_behavior casing;
     const char *file_search_string;
     int match_files;
