+++ src/print.c
@@ -34,19 +34,19 @@ void print_binary_file_matches(const char *path) {
     fprintf(out_fd, "Binary file %s matches.\n", path);
 }
 
-void print_file_matches(const char *path, const char *buf, const int buf_len, const match matches[], const int matches_len) {
-    int line = 1;
+void print_file_matches(const char *path, const char *buf, const size_t buf_len, const match matches[], const size_t matches_len) {
+    size_t line = 1;
     char **context_prev_lines = NULL;
-    int prev_line = 0;
-    int last_prev_line = 0;
-    int prev_line_offset = 0;
-    int cur_match = 0;
+    size_t prev_line = 0;
+    size_t last_prev_line = 0;
+    size_t prev_line_offset = 0;
+    size_t cur_match = 0;
     /* TODO the line below contains a terrible hack */
-    int lines_since_last_match = 1000000; /* if I initialize this to INT_MAX it'll overflow */
-    int lines_to_print = 0;
-    int last_printed_match = 0;
+    size_t lines_since_last_match = 1000000; /* if I initialize this to INT_MAX it'll overflow */
+    ssize_t lines_to_print = 0;
+    size_t last_printed_match = 0;
     char sep = '-';
-    int i, j;
+    size_t i, j;
     int in_a_match = FALSE;
     int printing_a_match = FALSE;
 
@@ -82,7 +82,7 @@ void print_file_matches(const char *path, const char *buf, const int buf_len, co
                 lines_to_print = lines_since_last_match - (opts.after + 1);
                 if (lines_to_print < 0) {
                     lines_to_print = 0;
-                } else if (lines_to_print > opts.before) {
+                } else if ((size_t)lines_to_print > opts.before) {
                     lines_to_print = opts.before;
                 }
 
@@ -121,7 +121,7 @@ void print_file_matches(const char *path, const char *buf, const int buf_len, co
                     /* print headers for ackmate to parse */
                     print_line_number(line, ';');
                     for (; last_printed_match < cur_match; last_printed_match++) {
-                        fprintf(out_fd, "%i %i",
+                        fprintf(out_fd, "%lu %lu",
                                 (matches[last_printed_match].start - prev_line_offset),
                                 (matches[last_printed_match].end - matches[last_printed_match].start));
                         last_printed_match == cur_match - 1 ? fputc(':', out_fd) : fputc(',', out_fd);
@@ -134,7 +134,7 @@ void print_file_matches(const char *path, const char *buf, const int buf_len, co
                 } else {
                     print_line_number(line, ':');
                     if (opts.column) {
-                        fprintf(out_fd, "%i:", (matches[last_printed_match].start - prev_line_offset) + 1);
+                        fprintf(out_fd, "%lu:", (matches[last_printed_match].start - prev_line_offset) + 1);
                     }
 
                     if (printing_a_match && opts.color) {
@@ -196,7 +196,7 @@ void print_file_matches(const char *path, const char *buf, const int buf_len, co
     free(context_prev_lines);
 }
 
-void print_line_number(int line, const char sep) {
+void print_line_number(size_t line, const char sep) {
     if (!opts.print_line_numbers) {
         return;
     }
@@ -208,9 +208,9 @@ void print_line_number(int line, const char sep) {
     }
 
     if (opts.color) {
-        fprintf(out_fd, "%s%i%s%c", opts.color_line_number, line, color_reset, sep);
+        fprintf(out_fd, "%s%lu%s%c", opts.color_line_number, line, color_reset, sep);
     } else {
-        fprintf(out_fd, "%i%c", line, sep);
+        fprintf(out_fd, "%lu%c", line, sep);
     }
 }
